+++ packages/test-tube/src/runner/mod.rs
@@ -1,7 +1,7 @@
 use cosmwasm_std::CosmosMsg;
 
 use crate::account::SigningAccount;
-use crate::runner::result::{RunnerExecuteResult, RunnerResult};
+use crate::runner::result::{RunnerExecuteResult, RunnerExecuteResultMult, RunnerResult};
 use crate::utils::{bank_msg_to_any, wasm_msg_to_any};
 use crate::RunnerError;
 
@@ -28,6 +28,14 @@ pub trait Runner<'a> {
         msgs: &[(M, &str)],
         signer: &SigningAccount,
     ) -> RunnerExecuteResult<R>
+    where
+        M: ::prost::Message,
+        R: ::prost::Message + Default;
+
+    fn execute_single_block<M, R>(
+        &self,
+        msgs: &[(M, &str, &SigningAccount)],
+    ) -> RunnerExecuteResultMult<R>
     where
         M: ::prost::Message,
         R: ::prost::Message + Default;